requirement.txt
fastapi
uvicorn
motor
python-dotenv
python-jose
passlib[bcrypt]
pydantic